+++ CHANGELOG.md
@@ -5,6 +5,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file follows the convention described at
 [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+### Changed
+- Removed unnecessary use of `Box` from `StreamCDC` in `v2016` and `v2020`.
+
 ## [3.0.1] - 2023-02-28
 ### Added
 - nagy: Support conversion to `std::io::Error` in streaming chunkers.